Added reverse strand allele correction & command line execution
Reverse strand allele correction:
If variants are mapped onto the reverse strand of the new assembly, the REF alleles are now ajusted accordingly, (example: the REF allele A becomes T) so that the VCF file always describes variants for the forward strand.
Command line execution:
The pipeline can now be executed via command line, see README for instructions.
Additional features:
Added calculation of the % of remapped variants (output in console).